refactor(lottoNubmers): delete lottoNumbers
