+++ src/test/java/LottoTest.java
@@ -3,7 +3,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import domain.Lotto;
 import domain.LottoNumber;
-import java.util.List;
+import java.util.ArrayList;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
@@ -30,35 +30,37 @@ public class LottoTest {
             .hasMessageContaining("[ERROR] 로또 번호로 null 값이 올 수 없습니다.");
     }
 
-//
-//    @Test
-//    @DisplayName("로또 숫자 1~45 사이 아닐 시 예외 발생")
-//    void check_range_fail() {
-//
-//        assertThatThrownBy(
-//            () -> new Lotto(List.of(1, 2, 3, 4, 5, 46))
-//        ).isInstanceOf(IllegalArgumentException.class)
-//            .hasMessageContaining("[ERROR] 로또 번호는 1~45 사이 정수만 가능합니다.");
-//
-//    }
-//
-//    @Test
-//    @DisplayName("로또 숫자가 6개가 아닐 시 예외 발생")
-//    void check_size_fail() {
-//        assertThatThrownBy(
-//            () -> new Lotto(List.of(1, 2, 3, 4, 5, 6, 7))
-//        ).isInstanceOf(IllegalArgumentException.class)
-//            .hasMessageContaining("[ERROR] 로또 번호는 6개만 입력 가능합니다.");
-//    }
-//
-//    @Test
-//    @DisplayName("로또 번호에 중복이 있을 시 예외 발생")
-//    void check_unique_fail() {
-//        assertThatThrownBy(
-//            () -> new Lotto(List.of(1, 2, 3, 4, 5, 5))
-//        ).isInstanceOf(IllegalArgumentException.class)
-//            .hasMessageContaining("[ERROR] 로또 번호는 중복될 수 없습니다.");
-//    }
+    @Test
+    @DisplayName("로또 번호에 빈 값이 올 시 예외 발생")
+    void lotto_numbers_empty_fail() {
+
+        assertThatThrownBy(() -> new Lotto(new ArrayList<>()))
+            .isInstanceOf(IllegalArgumentException.class)
+            .hasMessageContaining("[ERROR] 로또 번호는 6개만 입력 가능합니다.");
+
+    }
+
+    @Test
+    @DisplayName("로또 숫자가 6개가 아닐 시 예외 발생")
+    void check_size_fail() {
+        assertThatThrownBy(
+            () -> new Lotto(Stream.of(1, 2, 3, 4, 5, 6, 7)
+                .map(LottoNumber::new)
+                .collect(Collectors.toList()))
+        ).isInstanceOf(IllegalArgumentException.class)
+            .hasMessageContaining("[ERROR] 로또 번호는 6개만 입력 가능합니다.");
+    }
+
+    @Test
+    @DisplayName("로또 번호에 중복이 있을 시 예외 발생")
+    void check_unique_fail() {
+        assertThatThrownBy(
+            () -> new Lotto(Stream.of(1, 2, 3, 4, 5, 5)
+                .map(LottoNumber::new)
+                .collect(Collectors.toList()))
+        ).isInstanceOf(IllegalArgumentException.class)
+            .hasMessageContaining("[ERROR] 로또 번호는 중복될 수 없습니다.");
+    }
 
 
 }
